refactor: moved UI library to shared/component folder
