+++ src/app/(auth)/login/page.tsx
@@ -3,7 +3,7 @@ import { useRouter } from "next/navigation";
 import { signIn } from "next-auth/react";
 import { useState } from "react";
 
-import Button from "@/components/UI/button";
+import Button from "@/shared/components/UI/button";
 
 const Login = () => {
   const router = useRouter();
