+++ src/app/(store)/list/[[...params]]/page.tsx
@@ -9,13 +9,13 @@ import { getList } from "@/actions/list/listServices";
 import ProductCard from "@/components/store/common/productCard";
 import Filters from "@/components/store/listPage/filters";
 import NoItem from "@/components/store/listPage/noItem";
-import Button from "@/components/UI/button";
-import DropDownList from "@/components/UI/dropDown";
-import LineList from "@/components/UI/lineList";
 import { ProductListSkeleton } from "@/domains/store/productList/components";
 import { DEFAULT_FILTERS, SORT_DATA, sortDropdownData } from "@/domains/store/productList/constants";
 import { TFilterBrands, TFilters, TListItem } from "@/domains/store/productList/types";
 import { getFiltersFromProductList } from "@/domains/store/productList/utils";
+import Button from "@/shared/components/UI/button";
+import DropDownList from "@/shared/components/UI/dropDown";
+import LineList from "@/shared/components/UI/lineList";
 import { IMAGE_BASE_URL } from "@/shared/constants/store";
 import { cn } from "@/shared/utils/styling";
 import { TPageStatus } from "@/types/list";
