+++ src/app/admin/trafficView/[pageNumber]/page.tsx
@@ -4,9 +4,9 @@ import { useParams } from "next/navigation";
 import { useCallback, useEffect, useState } from "react";
 
 import { TTrafficListItem, deleteTraffic, getTrafficReport } from "@/actions/pageVisit/pageVisitServices";
-import Button from "@/components/UI/button";
-import { SK_Box } from "@/components/UI/skeleton";
-import { Pagination } from "@/components/UI/table/pagination";
+import Button from "@/shared/components/UI/button";
+import { SK_Box } from "@/shared/components/UI/skeleton";
+import { Pagination } from "@/shared/components/UI/table/pagination";
 import { TRAFFIC_LIST_PAGE_SIZE } from "@/shared/constants/admin/trafficView";
 import { getFullTimeString } from "@/shared/utils/formatting/time";
 import { cn } from "@/shared/utils/styling";
