+++ src/components/admin/category/addCategoryGroup/index.tsx
@@ -2,8 +2,8 @@
 import { useState } from "react";
 
 import { TGetAllCategories, addCategory } from "@/actions/category/category";
-import Button from "@/components/UI/button";
-import Popup from "@/components/UI/popup";
+import Button from "@/shared/components/UI/button";
+import Popup from "@/shared/components/UI/popup";
 
 import GroupCategory from "../../forms/groupCategory";
