+++ src/components/admin/category/rowGroup/addCategory/index.tsx
@@ -1,6 +1,6 @@
 "use client";
 import { TAddCategory } from "@/actions/category/category";
-import Input from "@/components/UI/input";
+import Input from "@/shared/components/UI/input";
 
 type TProps = {
   errorMsg: string;
