+++ src/components/admin/category/rowGroup/category/index.tsx
@@ -10,8 +10,8 @@ import {
   TGetAllCategories,
   TUpdateCategory,
 } from "@/actions/category/category";
-import Button from "@/components/UI/button";
-import Popup from "@/components/UI/popup";
+import Button from "@/shared/components/UI/button";
+import Popup from "@/shared/components/UI/popup";
 
 import AddCategory from "../addCategory";
 import CategoryOptions from "../categoryOptions";
