+++ src/components/admin/product/productForm/index.tsx
@@ -6,9 +6,9 @@ import { useEffect, useState } from "react";
 import { getAllBrands } from "@/actions/brands/brands";
 import { getAllCategoriesJSON } from "@/actions/category/category";
 import { getCategorySpecs } from "@/actions/category/specifications";
-import Button from "@/components/UI/button";
-import DropDownList from "@/components/UI/dropDown";
-import Input from "@/components/UI/input";
+import Button from "@/shared/components/UI/button";
+import DropDownList from "@/shared/components/UI/dropDown";
+import Input from "@/shared/components/UI/input";
 import { TBrand } from "@/shared/types";
 import { cn } from "@/shared/utils/styling";
 import { TGroupJSON } from "@/types/categories";
