+++ src/components/store/common/shoppingCart/index.tsx
@@ -5,14 +5,13 @@ import { useSelector } from "react-redux";
 
 import { getCartProducts } from "@/actions/product/product";
 import { CloseIcon, ShoppingIconEmpty } from "@/components/icons/svgIcons";
-import Button from "@/components/UI/button";
+import Button from "@/shared/components/UI/button";
 import { cn } from "@/shared/utils/styling";
 import { RootState } from "@/store/shoppingCart";
 import { TCartListItemDB } from "@/types/product";
 import { TCartItemData } from "@/types/shoppingCart";
 
 import CartItem from "./_components/cartItem";
-
 
 type TProps = {
   isVisible: boolean;
