+++ src/components/store/navbar/navProfile/index.tsx
@@ -3,7 +3,7 @@
 import { useRef } from "react";
 
 import { ProfileIcon } from "@/components/icons/svgIcons";
-import Button from "@/components/UI/button";
+import Button from "@/shared/components/UI/button";
 import { useToggleMenu } from "@/shared/hooks/useToggleMenu";
 import { cn } from "@/shared/utils/styling";
