+++ src/shared/components/UI/checkBox/index.tsx
