+++ src/shared/components/UI/priceSlider/index.tsx
