+++ src/shared/components/UI/skeleton/index.tsx
